Update hasArgument to conform to AsIs traversal.
Since [1], the hasArgument matcher has ignored implicit casts and
parenthesis when matching the argument expression. However, with the
introduction of traversal mode IgnoreUnlessSpelledInSource, the behavior
was updated in [2] to disregard defaulted arguments. Nevertheless, when
using AsIs, it would be expected that the matched expression includes
implicit nodes. Furthermore, even with IgnoreUnlessSpelledInSource, it
no longer makes sense to skip parenthesized arguments. Therefore, this
change aims to update the behavior to better fall in line with the
expectations of the recently introuduced traversal modes.

[1] https://reviews.llvm.org/rG04616e47769af372649ec2c16e89a2fbc1ddce8b
[2] https://reviews.llvm.org/D90984

FIX: breaks (at least) clang-tidys relying on the prev implementation.